Chore: Move common function to common conftest
